requirements.txt
Flask==2.3.3
flask-cors==6.0.1
PyJWT==2.10.1
cryptography==45.0.6
google-auth==2.40.3
together==1.5.17
tavily-python==0.7.12
requests==2.32.4
tiktoken>=0.7.0
pydantic==2.11.7
python-dotenv==1.1.1
boto3 
gunicorn>=22.1
psycopg2-binary
python-magic==0.4.27
pypdf==5.9.0
python-docx==1.1.2
openpyxl==3.1.5
gevent>=24.0.0
flask-socketio>=5.3.0
python-socketio>=5.10.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0

